html/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
gunicorn>=23.0.0
asyncpg>=0.30.0
pydantic-settings>=2.6.0
PyJWT>=2.10.0
bcrypt>=4.2.0
jinja2>=3.1.0
python-multipart>=0.0.20
aiofiles>=24.1.0
prometheus-fastapi-instrumentator>=7.0.0
python-json-logger>=3.0.0
